requirements.txt
python-telegram-bot==20.7
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
flask==3.0.0
requests==2.31.0
yt-dlp==2023.11.16
pydub==0.25.1
moviepy==1.0.3
psycopg2-binary==2.9.9
gunicorn==21.2.0